* filer/spam.rb,plugin/10spamfilter.rb: fix spamfilter.max_rate float to interger.
git-svn-id: https://tdiary.svn.sourceforge.net/svnroot/tdiary/trunk/core@2885 7f22e88f-374d-0410-998f-c91420d97ba2

+++ tdiary/filter/spam.rb
@@ -51,7 +51,7 @@ module TDiary
 				end
 
 				if @conf.options.include?('spamfilter.max_rate')
-					@max_rate = @conf.options['spamfilter.max_rate'].to_f
+					@max_rate = @conf.options['spamfilter.max_rate'].to_i
 				else
 					@max_rate = nil
 				end
